fix(pr-135): align SQL type casts, add React imports, refactor practice handler
- Fixed persist_practice_schedule: id::uuid, season_settings_id::bigint
- Fixed persist_game_schedule: id::uuid, season_settings_id::bigint
- Fixed persist_team_schedule: season_settings_id::bigint
- Added useState import to GamePersistencePanel.jsx and PracticePersistencePanel.jsx
- Refactored practicePersistenceHandler.js to use handlePersistenceRequest + evaluateOverrides
- Updated PR description with code review fixes

+++ supabase/migrations/20251206000000_team_persistence_rpc.sql
@@ -30,7 +30,7 @@ BEGIN
     )
     SELECT
         (run_data->>'id')::uuid,
-        (run_data->>'season_settings_id')::int,
+        (run_data->>'season_settings_id')::bigint,
         run_data->>'run_type',
         run_data->>'status',
         run_data->'parameters',
